Move quickstart.rb -> storage_quickstart.rb
