+++ google-cloud-storage/samples/storage_quickstart.rb
